abstract some logic away from the server and into a module, which is imported into the server.  also, changed the import type from common-js modules to es6 modules
