+++ app.js
@@ -1,29 +1,19 @@
-const express = require('express')
-const cors = require('cors')
+import express from 'express';
+import cors from 'cors'; 
+import {get, add} from './api/index.js'
 const app = express()
 const port = 3001
+
 
 // parses incoming to JSON to req.body
 app.use(cors()); 
 app.use(express.json());
 
-// defines array
-const people = [];
-
-// get request @localhost:3000/api
-app.get('/api', (req, res) => {
-  // returns your array in JSON format
-  res.json(people)
-})
-
+// get request @localhost:3000/api & return array in JSON format
+app.get('/api', (req, res) => get(req, res));
+  
 //recieves post request
-app.post('/api', (req, res) => {
-  console.log(req.body)
-  // adds body to array
-  people.push(req.body)
-  // return confirmation
-  res.json(req.body)
-})
+app.post('/api', (req, res) => add(req, res));
 
 // opens/ listen on port
 app.listen(port, () => {
