+++ package.json
@@ -6,6 +6,7 @@
   "name": "js-vs-code",
   "version": "1.0.0",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
